requirements.txt
alpaca-ml==0.8.2
transformers==3.1.0
scikit-learn==0.22.2.post1
hydra-core==0.11.3
numpy==1.18.2
wandb==0.10.27
datasets==1.4.0